+++ src/main/java/org/zalando/nakadi/exceptions/runtime/ServiceTemporarilyUnavailableException.java
@@ -5,6 +5,10 @@ public class ServiceTemporarilyUnavailableException extends MyNakadiRuntimeExcep
         super(cause.getMessage(), cause);
     }
 
+    public ServiceTemporarilyUnavailableException(final String message) {
+        super(message);
+    }
+
     public ServiceTemporarilyUnavailableException(final String message, final Throwable cause) {
         super(message, cause);
     }