+++ src/main/java/org/zalando/nakadi/service/subscription/SubscriptionService.java
@@ -15,6 +15,7 @@ import org.zalando.nakadi.domain.ItemsWrapper;
 import org.zalando.nakadi.domain.NakadiCursor;
 import org.zalando.nakadi.domain.PaginationLinks;
 import org.zalando.nakadi.domain.PaginationWrapper;
+import org.zalando.nakadi.domain.PartitionBaseStatistics;
 import org.zalando.nakadi.domain.PartitionEndStatistics;
 import org.zalando.nakadi.domain.Subscription;
 import org.zalando.nakadi.domain.SubscriptionBase;
@@ -24,7 +25,6 @@ import org.zalando.nakadi.exceptions.InternalNakadiException;
 import org.zalando.nakadi.exceptions.InvalidCursorException;
 import org.zalando.nakadi.exceptions.NoSuchEventTypeException;
 import org.zalando.nakadi.exceptions.NoSuchSubscriptionException;
-import org.zalando.nakadi.exceptions.ServiceUnavailableException;
 import org.zalando.nakadi.exceptions.Try;
 import org.zalando.nakadi.exceptions.runtime.DbWriteOperationsBlockedException;
 import org.zalando.nakadi.exceptions.runtime.DuplicatedSubscriptionException;
@@ -137,7 +137,7 @@ public class SubscriptionService {
     }
 
     public Result listSubscriptions(@Nullable final String owningApplication, @Nullable final Set<String> eventTypes,
-                                    final int limit, final int offset) {
+                                    final boolean showStatus, final int limit, final int offset) {
         if (limit < 1 || limit > 1000) {
             final Problem problem = Problem.valueOf(Response.Status.BAD_REQUEST,
                     "'limit' parameter should have value from 1 to 1000");
@@ -156,11 +156,17 @@ public class SubscriptionService {
             final List<Subscription> subscriptions =
                     subscriptionRepository.listSubscriptions(eventTypesFilter, owningAppOption, offset, limit);
             final PaginationLinks paginationLinks = SubscriptionsUriHelper.createSubscriptionPaginationLinks(
-                    owningAppOption, eventTypesFilter, offset, limit, subscriptions.size());
-            return Result.ok(new PaginationWrapper<>(subscriptions, paginationLinks));
-        } catch (final ServiceUnavailableException e) {
+                    owningAppOption, eventTypesFilter, offset, limit, showStatus, subscriptions.size());
+            final PaginationWrapper<Subscription> paginationWrapper =
+                    new PaginationWrapper<>(subscriptions, paginationLinks);
+            if (showStatus) {
+                final List<Subscription> items = paginationWrapper.getItems();
+                items.forEach(s -> s.setStatus(createSubscriptionStat(s, false)));
+            }
+            return Result.ok(paginationWrapper);
+        } catch (final ServiceTemporarilyUnavailableException e) {
             LOG.error("Error occurred during listing of subscriptions", e);
-            return Result.problem(e.asProblem());
+            return Result.problem(Problem.valueOf(Response.Status.SERVICE_UNAVAILABLE, e.getMessage()));
         }
     }
 
@@ -171,9 +177,9 @@ public class SubscriptionService {
         } catch (final NoSuchSubscriptionException e) {
             LOG.debug("Failed to find subscription: {}", subscriptionId);
             return Result.problem(e.asProblem());
-        } catch (final ServiceUnavailableException e) {
+        } catch (final ServiceTemporarilyUnavailableException e) {
             LOG.error("Error occurred when trying to get subscription: {}", subscriptionId, e);
-            return Result.problem(e.asProblem());
+            return Result.problem(Problem.valueOf(Response.Status.SERVICE_UNAVAILABLE, e.getMessage()));
         }
     }
 
@@ -198,64 +204,65 @@ public class SubscriptionService {
         } catch (final NoSuchSubscriptionException e) {
             LOG.debug("Failed to find subscription: {}", subscriptionId, e);
             return Result.problem(e.asProblem());
-        } catch (final ServiceUnavailableException e) {
+        } catch (final ServiceTemporarilyUnavailableException e) {
             LOG.error("Error occurred when trying to delete subscription: {}", subscriptionId, e);
-            return Result.problem(e.asProblem());
+            return Result.problem(Problem.valueOf(Response.Status.SERVICE_UNAVAILABLE, e.getMessage()));
         } catch (final NoSuchEventTypeException | InternalNakadiException e) {
             LOG.error("Exception can not occur", e);
             return Result.problem(e.asProblem());
         }
     }
 
-    public ItemsWrapper<SubscriptionEventTypeStats> getSubscriptionStat(final String subscriptionId)
+    public ItemsWrapper<SubscriptionEventTypeStats> getSubscriptionStat(final String subscriptionId,
+                                                                        final boolean includeDistance)
             throws InconsistentStateException, NoSuchSubscriptionException, ServiceTemporarilyUnavailableException {
         final Subscription subscription;
         try {
             subscription = subscriptionRepository.getSubscription(subscriptionId);
-        } catch (final ServiceUnavailableException ex) {
+        } catch (final ServiceTemporarilyUnavailableException ex) {
             throw new InconsistentStateException(ex.getMessage());
         }
-        final List<SubscriptionEventTypeStats> subscriptionStat = createSubscriptionStat(subscription);
+        final List<SubscriptionEventTypeStats> subscriptionStat = createSubscriptionStat(subscription, includeDistance);
         return new ItemsWrapper<>(subscriptionStat);
     }
 
-
-    private List<SubscriptionEventTypeStats> createSubscriptionStat(final Subscription subscription)
+    private List<SubscriptionEventTypeStats> createSubscriptionStat(final Subscription subscription,
+                                                                    final boolean includeDistance)
             throws InconsistentStateException, ServiceTemporarilyUnavailableException {
+        final List<EventType> eventTypes = getEventTypesForSubscription(subscription);
+        final ZkSubscriptionClient subscriptionClient = createZkSubscriptionClient(subscription);
+        final Optional<ZkSubscriptionNode> zkSubscriptionNode = subscriptionClient.getZkSubscriptionNode();
 
-        final List<EventType> eventTypes = subscription.getEventTypes().stream()
+        if (includeDistance) {
+            return loadStats(eventTypes, zkSubscriptionNode, subscriptionClient);
+        } else {
+            return loadLightStats(eventTypes, zkSubscriptionNode);
+        }
+    }
+
+    private ZkSubscriptionClient createZkSubscriptionClient(final Subscription subscription)
+            throws ServiceTemporarilyUnavailableException {
+        try {
+            return subscriptionClientFactory.createClient(subscription,
+                    "subscription." + subscription.getId() + ".stats");
+        } catch (final InternalNakadiException | NoSuchEventTypeException e) {
+            throw new ServiceTemporarilyUnavailableException(e);
+        }
+    }
+
+    private List<EventType> getEventTypesForSubscription(final Subscription subscription) {
+        return subscription.getEventTypes().stream()
                 .map(Try.wrap(eventTypeRepository::findByName))
                 .map(Try::getOrThrow)
                 .sorted(Comparator.comparing(EventType::getName))
                 .collect(Collectors.toList());
-
-        final List<PartitionEndStatistics> topicPartitions;
-        try {
-            topicPartitions = loadPartitionEndStatistics(eventTypes);
-        } catch (final ServiceUnavailableException ex) {
-            throw new ServiceTemporarilyUnavailableException(ex);
-        }
-
-        final ZkSubscriptionClient subscriptionClient;
-        try {
-            subscriptionClient = subscriptionClientFactory.createClient(
-                    subscription, "subscription." + subscription.getId() + ".stats");
-        } catch (final InternalNakadiException | NoSuchEventTypeException e) {
-            throw new ServiceTemporarilyUnavailableException(e);
-        }
-
-        final Optional<ZkSubscriptionNode> zkSubscriptionNode = subscriptionClient.getZkSubscriptionNodeLocked();
-
-        return loadStats(eventTypes, zkSubscriptionNode, subscriptionClient, topicPartitions);
     }
 
     private List<PartitionEndStatistics> loadPartitionEndStatistics(final Collection<EventType> eventTypes)
-            throws ServiceUnavailableException {
+            throws ServiceTemporarilyUnavailableException {
         final List<PartitionEndStatistics> topicPartitions = new ArrayList<>();
 
-        final Map<TopicRepository, List<Timeline>> timelinesByRepo = eventTypes.stream()
-                .map(timelineService::getActiveTimeline)
-                .collect(Collectors.groupingBy(timelineService::getTopicRepository));
+        final Map<TopicRepository, List<Timeline>> timelinesByRepo = getTimelinesByRepository(eventTypes);
 
         for (final Map.Entry<TopicRepository, List<Timeline>> repoEntry : timelinesByRepo.entrySet()) {
             final TopicRepository topicRepository = repoEntry.getKey();
@@ -265,58 +272,122 @@ public class SubscriptionService {
         return topicPartitions;
     }
 
+    private Map<TopicRepository, List<Timeline>> getTimelinesByRepository(final Collection<EventType> eventTypes) {
+        return eventTypes.stream()
+                .map(timelineService::getActiveTimeline)
+                .collect(Collectors.groupingBy(timelineService::getTopicRepository));
+    }
+
+    private List<String> getPartitionsList(final EventType eventType) {
+        final Timeline activeTimeline = timelineService.getActiveTimeline(eventType);
+        return timelineService.getTopicRepository(activeTimeline).listPartitionNames(activeTimeline.getTopic());
+    }
+
     private List<SubscriptionEventTypeStats> loadStats(
             final Collection<EventType> eventTypes,
             final Optional<ZkSubscriptionNode> subscriptionNode,
-            final ZkSubscriptionClient client,
-            final List<PartitionEndStatistics> stats)
+            final ZkSubscriptionClient client)
             throws ServiceTemporarilyUnavailableException, InconsistentStateException {
         final List<SubscriptionEventTypeStats> result = new ArrayList<>(eventTypes.size());
-
-        final Collection<NakadiCursor> committedPositions = subscriptionNode
-                .map(node -> loadCommittedPositions(node.getPartitions(), client))
-                .orElse(Collections.emptyList());
-
+        final Collection<NakadiCursor> committedPositions = getCommittedPositions(subscriptionNode, client);
+        final List<PartitionEndStatistics> stats = loadPartitionEndStatistics(eventTypes);
         for (final EventType eventType : eventTypes) {
-            final List<SubscriptionEventTypeStats.Partition> resultPartitions = new ArrayList<>(stats.size());
-            for (final PartitionEndStatistics stat : stats) {
-                final NakadiCursor lastPosition = stat.getLast();
-                if (!lastPosition.getEventType().equals(eventType.getName())) {
-                    continue;
-                }
-                final Long distance = committedPositions.stream()
-                        .filter(pos -> pos.getEventTypePartition().equals(lastPosition.getEventTypePartition()))
-                        .findAny()
-                        .map(committed -> {
-                            try {
-                                return cursorOperationsService.calculateDistance(committed, lastPosition);
-                            } catch (final InvalidCursorOperation ex) {
-                                throw new InconsistentStateException(
-                                        "Unexpected exception while calculating distance", ex);
-                            }
-                        })
-                        .orElse(null);
-
-                final Partition.State state = subscriptionNode
-                        .map(node -> node.guessState(stat.getTimeline().getEventType(), stat.getPartition()))
-                        .orElse(Partition.State.UNASSIGNED);
-
-                final String streamId = subscriptionNode
-                        .map(node -> node.guessStream(stat.getTimeline().getEventType(), stat.getPartition()))
-                        .orElse("");
-
-                final SubscriptionEventTypeStats.Partition.AssignmentType assignmentType = subscriptionNode
-                        .map(node -> node.getPartitionAssignmentType(
-                                stat.getTimeline().getEventType(), stat.getPartition()))
-                        .orElse(null);
-
-                resultPartitions.add(new SubscriptionEventTypeStats.Partition(
-                        lastPosition.getPartition(), state.getDescription(), distance, streamId, assignmentType));
-            }
-            resultPartitions.sort(Comparator.comparing(SubscriptionEventTypeStats.Partition::getPartition));
-            result.add(new SubscriptionEventTypeStats(eventType.getName(), resultPartitions));
+            final List<PartitionBaseStatistics> statsForEventType = stats.stream()
+                    .filter(s -> s.getTimeline().getEventType().equals(eventType.getName()))
+                    .collect(Collectors.toList());
+            result.add(getEventTypeStats(subscriptionNode, eventType.getName(), statsForEventType,
+                    committedPositions));
         }
         return result;
+    }
+
+    private List<SubscriptionEventTypeStats> loadLightStats(final Collection<EventType> eventTypes,
+                                                            final Optional<ZkSubscriptionNode> subscriptionNode)
+            throws ServiceTemporarilyUnavailableException {
+        final List<SubscriptionEventTypeStats> result = new ArrayList<>(eventTypes.size());
+        for (final EventType eventType : eventTypes) {
+            result.add(getEventTypeLightStats(subscriptionNode, eventType));
+        }
+        return result;
+    }
+
+    private SubscriptionEventTypeStats getEventTypeStats(final Optional<ZkSubscriptionNode> subscriptionNode,
+                                                         final String eventTypeName,
+                                                         final List<? extends PartitionBaseStatistics> stats,
+                                                         final Collection<NakadiCursor> committedPositions) {
+        final List<SubscriptionEventTypeStats.Partition> resultPartitions =
+                new ArrayList<>(stats.size());
+        for (final PartitionBaseStatistics stat : stats) {
+            final String partition = stat.getPartition();
+            final NakadiCursor lastPosition = ((PartitionEndStatistics) stat).getLast();
+            final Long distance = computeDistance(committedPositions, lastPosition);
+            resultPartitions.add(getPartitionStats(subscriptionNode, eventTypeName, partition, distance));
+        }
+        resultPartitions.sort(Comparator.comparing(SubscriptionEventTypeStats.Partition::getPartition));
+        return new SubscriptionEventTypeStats(eventTypeName, resultPartitions);
+    }
+
+    private SubscriptionEventTypeStats getEventTypeLightStats(final Optional<ZkSubscriptionNode> subscriptionNode,
+                                                              final EventType eventType) {
+        final List<SubscriptionEventTypeStats.Partition> resultPartitions = new ArrayList<>();
+        for (final String partition: getPartitionsList(eventType)) {
+            resultPartitions.add(getPartitionStats(subscriptionNode, eventType.getName(), partition, null));
+        }
+        resultPartitions.sort(Comparator.comparing(SubscriptionEventTypeStats.Partition::getPartition));
+        return new SubscriptionEventTypeStats(eventType.getName(), resultPartitions);
+    }
+
+    private SubscriptionEventTypeStats.Partition getPartitionStats(final Optional<ZkSubscriptionNode> subscriptionNode,
+                                                                   final String eventTypeName, final String partition,
+                                                                   final Long distance) {
+        final Partition.State state = getState(subscriptionNode, eventTypeName, partition);
+        final String streamId = getStreamId(subscriptionNode, eventTypeName, partition);
+        final SubscriptionEventTypeStats.Partition.AssignmentType assignmentType =
+                getAssignmentType(subscriptionNode, eventTypeName, partition);
+        return new SubscriptionEventTypeStats.Partition(partition, state.getDescription(),
+                distance, streamId, assignmentType);
+    }
+
+    private Collection<NakadiCursor> getCommittedPositions(final Optional<ZkSubscriptionNode> subscriptionNode,
+                                                           final ZkSubscriptionClient client) {
+        return subscriptionNode
+                .map(node -> loadCommittedPositions(node.getPartitions(), client))
+                .orElse(Collections.emptyList());
+    }
+
+    private Partition.State getState(final Optional<ZkSubscriptionNode> subscriptionNode, final String eventType,
+                                     final String partition) {
+        return subscriptionNode.map(node -> node.guessState(eventType, partition))
+                .orElse(Partition.State.UNASSIGNED);
+    }
+
+    private SubscriptionEventTypeStats.Partition.AssignmentType getAssignmentType(
+            final Optional<ZkSubscriptionNode> subscriptionNode, final String eventType, final String partition) {
+        return subscriptionNode
+                .map(node -> node.getPartitionAssignmentType(eventType, partition))
+                .orElse(null);
+    }
+
+    private String getStreamId(final Optional<ZkSubscriptionNode> subscriptionNode,
+                               final String eventType, final String partition) {
+        return subscriptionNode
+                .map(node -> node.guessStream(eventType, partition))
+                .orElse("");
+    }
+
+    private Long computeDistance(final Collection<NakadiCursor> committedPositions, final NakadiCursor lastPosition) {
+        return committedPositions.stream()
+                .filter(pos -> pos.getEventTypePartition().equals(lastPosition.getEventTypePartition()))
+                .findAny()
+                .map(committed -> {
+                    try {
+                        return cursorOperationsService.calculateDistance(committed, lastPosition);
+                    } catch (final InvalidCursorOperation ex) {
+                        throw new InconsistentStateException(
+                                "Unexpected exception while calculating distance", ex);
+                    }
+                })
+                .orElse(null);
     }
 
     private Collection<NakadiCursor> loadCommittedPositions(
@@ -328,8 +399,7 @@ public class SubscriptionService {
                     partitions.stream().map(Partition::getKey).collect(Collectors.toList()));
 
             return converter.convert(committed.values());
-        } catch (final InternalNakadiException | NoSuchEventTypeException | InvalidCursorException |
-                ServiceUnavailableException e) {
+        } catch (final InternalNakadiException | NoSuchEventTypeException | InvalidCursorException e) {
             throw new ServiceTemporarilyUnavailableException(e);
         }
     }