+++ CHANGELOG.md
@@ -6,8 +6,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+## [2.6.4] - 2018-04-25
+
+### Added
+- Add optional status to the /subscriptions endpoint
+
 ### Fixed
+- Fixed commit for subscriptions that use direct assignment of partitions
 - Fixed OutOfMemoryError when using huge values for batch_limit and max_uncommitted_events
+
 
 ## [2.6.3] - 2018-04-10